move infinity-related files to their own folders
makes the folder structure a bit cleaner
