+++ src/core/infinity/infinity-upgrades.js
@@ -1,5 +1,5 @@
-import { GameMechanicState, SetPurchasableMechanicState } from "./game-mechanics";
-import { DC } from "./constants";
+import { GameMechanicState, SetPurchasableMechanicState } from "../game-mechanics";
+import { DC } from "../constants";
 
 class ChargedInfinityUpgradeState extends GameMechanicState {
   constructor(config, upgrade) {
